### 2023-09-28 #5888 @danielskatz
👋 @Ig-dolci - thanks for your submission.  I see a couple of small issues:

1. The title is grammatically incorrect. incremental is an adverb, and in a place where a verb (perhaps a gerund) would normally go. Would "schedules to increment adjoint simulations" or "schedules for incrementing adjoint simulations" be a reasonable replacement?
2. There's a DOI that isn't valid (i.e., https://doi.org/10.1137/20M132620 does not resolve).

Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
