### 2023-09-28 #5888 @danielskatz
@lg-dolci - I think you also need to add a bit to your readme and overall documentation. Note that reviewers are going to be looking at these review criteria: https://joss.readthedocs.io/en/latest/review_criteria.html and your software (and paper) will have to meet them
