+++ docs/paper/paper.bib
@@ -30,7 +30,7 @@
   pages = {A1555--A1582},
   year = {2021},
   publisher = {SIAM},
-  doi = {10.1137/20M132620},
+  doi = {10.1137/20M1326209},
 }
 
 @article{Jansen2011,
