### 2023-09-29 #5888 @diehlpk
Hi @Ig-dolci can you please elaborate on this

`The complete documentation for checkpoint_schedules is available at https://www.firedrakeproject.org/checkpoint_schedules/.`

Is your code a plugin for firedrake?
