### 2023-09-29 #5888 @diehlpk
Hi @Ig-dolci another question is how does your software relates to this paper?

https://github.com/openjournals/joss-reviews/issues/1292
