### 2023-09-29 #5888 @Ig-dolci
> Is your code a plugin for firedrake?

Hi @diehlpk. I will incorporate more details into the README as soon as possible. Thank you.
