### 2023-09-29 #5888 @Ig-dolci
> Hi @Ig-dolci I briefly checked the repo and the [Readme.md](https://github.com/firedrakeproject/checkpoint_schedules/blob/main/README.md) is very brief and tells not much about the software. Please provide more details about the software.

I will incorporate more details into the README as soon as possible. Thank you.
