### 2023-09-29 #5888 @dham
> Hi @Ig-dolci another question is how does your software relates to this paper?
>
> #1292

That paper documents the reimplementation of dolfin-adjoint in Python by Sebastian Mitusch and Simon Funke. The original dolfin-adjoint had a limited capability to do incremental in-memory checkpointing using only the original Revolve algorithm, but that capability was lost in the rewrite. Reintroducing a much more capable incremental checkpointing capability in dolfin-adjoint (now really called pyadjoint) is a motivation for this work.

However, this work is not itself that reimplementation. We observed that the sequence of actions that make up a checkpointing schedule is model independent. We therefore built a small API for these actions (Forward, Reverse, Move, Copy, EndForward, EndReverse) and then teamed up with the developers of several checkpointing schedule algorithms (most recently h-revolve) to provide a library of classical and recent algorithms all accessible via the same API. James Maddison, who was involved in that effort and has also authored another checkpointing algorithm, was involved in that work and has adopted this work as the source of checkpointing schedules for his tlm-adjoint framework. We are also in the process of adopting checkpoint_schedules in pyadjoint.

The relationship to Firedrake is that both tlm-adjoint and pyadjoint can be employed to provide automatic adjoint capabilities to models written in Firedrake. However, checkpoint_schedules is not tied to Firedrake and its integration into both tlm-adjoint and pyadjoint demonstrates that it is sufficiently general that different frameworks can use it.
