### 2023-09-29 #5888 @diehlpk
Hi @Ig-dolci few more comments before we can start the review:

1. Is there any documentation on how to use the code?
2. Is there any documentation on how to install the code?

This might be added to the Readme.md.

Is there any link to the documentation?

3. Are there any guidelines for how to contribute or community guidelines?

From a first check, I would recommend that you read the [reviewer criteria](https://joss.readthedocs.io/en/latest/review_criteria.html)  and see if your code has these requirements.

In the current form, I think reviewers will raise many comments, and I am hesitated to send the code as it is to the reviewers.
