### 2023-10-02 5eff321
adding more details at README
