+++ README.md
@@ -1,5 +1,28 @@
 # checkpoint_schedules
-*checkpointing_schedules* provides schedules for step-based incremental
-checkpointing of the adjoints to computer models. 
+*checkpointing_schedules* provides schedules for step based incremental
+checkpointing of the adjoints to computer models. The schedules contain
+instructions indicating the sequence of forward and adjoint steps to be
+executed, and the data storage and retrieval to be performed. 
 
-The complete documentation for *checkpoint_schedules* is available at https://www.firedrakeproject.org/checkpoint_schedules/.
+The schedules instructions are independent of the checkpointing model implementation,
+which enables the model authors to switch between checkpointing algorithms
+without recoding their adjoint solvers. Conversely, *checkpointing_schedules*
+provides developers of checkpointing algorithms a direct mechanism to convey
+their work to model authors.
+
+## Installation
+*checkpointing_schedules* is a Python package and can be installed using pip
+```
+pip install checkpoint-schedules
+```
+
+## Usage
+The usage guide of *checkpoint_schedules* is available [here](https://nbviewer.org/github/firedrakeproject/checkpoint_schedules/blob/main/docs/notebooks/tutorial.ipynb).
+
+
+## Documentation
+The complete documentation for checkpoint_schedules is available at [Firedrake project website](https://www.firedrakeproject.org/checkpoint_schedules/).
+
+
+## License
+*checkpoint_schedules* is licensed under the GNU LGPL version 3. See the LICENSE file for details.
