### 2023-10-03 ccba0a2
Update README.md
checkpointing_schedules -> checkpoint_schedules
