+++ README.md
@@ -20,9 +20,18 @@ pip install checkpoint-schedules
 The usage guide of *checkpoint_schedules* is available [here](https://nbviewer.org/github/firedrakeproject/checkpoint_schedules/blob/main/docs/notebooks/tutorial.ipynb).
 
 
+## Contributing
+We welcome contributions to *checkpoint_schedules*!
+To contribute please consider the following steps:
+1. Fork the repository.
+2. Make your changes.
+3. Submit a pull request.
+
+## Bug reports
+Please report bugs on the [issue tracker](https://github.com/firedrakeproject/checkpoint_schedules/issues).
+
 ## Documentation
 The complete documentation for checkpoint_schedules is available at [Firedrake project website](https://www.firedrakeproject.org/checkpoint_schedules/).
-
 
 ## License
 *checkpoint_schedules* is licensed under the GNU LGPL version 3. See the LICENSE file for details.
