### 2023-10-04 eb0438e
Bugfix to SingleDiskStorageSchedule, update associated unit tests
