+++ tests/test_trivial.py
@@ -1,28 +1,29 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
+
 import functools
 import pytest
-from checkpoint_schedules.schedule import \
-    Forward, Reverse, Copy, Move, EndForward, EndReverse, StorageType
-from checkpoint_schedules import SingleDiskStorageSchedule, \
-    SingleMemoryStorageSchedule
+from checkpoint_schedules.schedule import (
+    Forward, Reverse, Copy, Move, EndForward, EndReverse, StorageType)
+from checkpoint_schedules import (
+    SingleDiskStorageSchedule, SingleMemoryStorageSchedule)
 
 
 def single_disk_copy(n):
-    checkpointing = SingleDiskStorageSchedule()
-    return (checkpointing,
+    cp_schedule = SingleDiskStorageSchedule(move_data=False)
+    return (cp_schedule,
             {StorageType.RAM: 0, StorageType.DISK: n}, 1)
 
 
 def single_disk_move(n):
-    checkpointing = SingleDiskStorageSchedule(move_data=True)
-    return (checkpointing,
+    cp_schedule = SingleDiskStorageSchedule(move_data=True)
+    return (cp_schedule,
             {StorageType.RAM: 0, StorageType.DISK: n}, 1)
 
 
 def single_memory(n):
-    checkpointing = SingleMemoryStorageSchedule()
-    return (checkpointing,
+    cp_schedule = SingleMemoryStorageSchedule()
+    return (cp_schedule,
             {StorageType.RAM: 0, StorageType.DISK: 0}, n)
 
 
@@ -35,17 +36,16 @@ def single_memory(n):
      ]
      )
 def test_validity(schedule, n=10):
-    """Test the checkpoint revolvers.
+    """Test basic checkpointing schedules.
 
     Parameters
     ----------
-    schedule : object
-        Revolver schedule.
+    schedule : callable
+        Accepts the number of forward steps and returns a schedule.
     n : int
-        Total forward steps.
-    S : int
-        Snapshots.
+        Number of forward steps
     """
+
     @functools.singledispatch
     def action(cp_action):
         raise TypeError("Unexpected action")
@@ -56,8 +56,6 @@ def test_validity(schedule, n=10):
         # Start at the current location of the forward
         assert model_n is not None and model_n == cp_action.n0
 
-        assert cp_action.storage == StorageType.WORK or \
-            cp_action.storage == StorageType.DISK
         # If the schedule has been finalized, end at or before the end of the
         # forward
         assert cp_schedule.max_n is None or cp_action.n1 <= n
@@ -68,21 +66,17 @@ def test_validity(schedule, n=10):
         n1 = min(cp_action.n1, n)
 
         model_n = n1
-        data.clear()
-        assert not cp_action.write_ics
-        assert cp_action.write_adj_deps is True
-        assert len(data.intersection(range(cp_action.n0, n1))) == 0
-        if cp_action.storage == StorageType.DISK:
-            for step in range(cp_action.n0, n1):
-                data.update(range(step, step + 1))
-                snapshots[cp_action.storage][step] = (set(ics), set(data))
-                data.clear()
-            assert cp_action.n0 == min(snapshots[cp_action.storage])
-        else:
-            data.update(range(cp_action.n0, n1))
-            assert cp_action.n0 == min(data)
-
         assert len(ics) == 0
+        assert not cp_action.write_ics
+        assert len(data.intersection(range(cp_action.n0, n1))) == 0
+        assert cp_action.write_adj_deps
+        if cp_action.storage == StorageType.DISK:
+            snapshots[cp_action.storage][cp_action.n0] = \
+                (set(ics), set(range(cp_action.n0, n1)))
[…]
