+++ tests/test_validity.py
@@ -19,23 +19,24 @@
 
 import functools
 import pytest
-from checkpoint_schedules.schedule import \
-    Forward, Reverse, Copy, Move, EndForward, EndReverse, StorageType
-from checkpoint_schedules import HRevolve, DiskRevolve, PeriodicDiskRevolve, \
-    Revolve, MultistageCheckpointSchedule, TwoLevelCheckpointSchedule, \
-    MixedCheckpointSchedule
+from checkpoint_schedules.schedule import (
+    Forward, Reverse, Copy, Move, EndForward, EndReverse, StorageType)
+from checkpoint_schedules import (
+    HRevolve, DiskRevolve, PeriodicDiskRevolve, Revolve,
+    MultistageCheckpointSchedule, TwoLevelCheckpointSchedule,
+    MixedCheckpointSchedule, SingleDiskStorageSchedule,
+    SingleMemoryStorageSchedule)
 
 
 def h_revolve(n, s):
-    snap_ram = s//3
-    snap_disk = s - s//3
-    if s//3 < 1:
-        pytest.skip("H-Revolve accepts snapshots in RAM > 1")
+    snap_ram = s // 3
+    snap_disk = s - snap_ram
+    if snap_ram < 1 or snap_disk < 1:
         return (None,
                 {StorageType.RAM: 0, StorageType.DISK: 0}, 0)
     else:
-        revolver = HRevolve(n, snap_ram, snap_disk)
-        return (revolver,
+        cp_schedule = HRevolve(n, snap_ram, snap_disk)
+        return (cp_schedule,
                 {StorageType.RAM: snap_ram, StorageType.DISK: snap_disk}, 1)
 
 
@@ -44,8 +45,8 @@ def disk_revolve(n, s):
         return (None,
                 {StorageType.RAM: 0, StorageType.DISK: 0}, 0)
     else:
-        revolver = DiskRevolve(n, s, n - s)
-        return (revolver,
+        cp_schedule = DiskRevolve(n, s, n - s)
+        return (cp_schedule,
                 {StorageType.RAM: s, StorageType.DISK: n - s}, 1)
 
 
@@ -64,9 +65,9 @@ def periodic_disk(n, s):
         return (None,
                 {StorageType.RAM: 0, StorageType.DISK: 0}, 0)
     else:
-        revolver = PeriodicDiskRevolve(n, s)
-        return (revolver,
-                {StorageType.RAM:  s, StorageType.DISK: n - s}, 1)
+        cp_schedule = PeriodicDiskRevolve(n, s)
+        return (cp_schedule,
+                {StorageType.RAM: s, StorageType.DISK: n - s}, 1)
 
 
 def revolve(n, s):
@@ -74,9 +75,9 @@ def revolve(n, s):
         return (None,
                 {StorageType.RAM: 0, StorageType.DISK: 0}, 0)
     else:
-        revolver = Revolve(n, s)
-        return (revolver,
-                {StorageType.RAM:  s, StorageType.DISK: 0}, 1)
+        cp_schedule = Revolve(n, s)
+        return (cp_schedule,
+                {StorageType.RAM: s, StorageType.DISK: 0}, 1)
 
 
 def mixed(n, s):
@@ -84,37 +85,63 @@ def mixed(n, s):
             {StorageType.RAM: 0, StorageType.DISK: s}, 1)
 
 
+def single_disk_copy(n, s):
+    cp_schedule = SingleDiskStorageSchedule(move_data=False)
+    return (cp_schedule,
+            {StorageType.RAM: 0, StorageType.DISK: n}, 1)
+
+
+def single_disk_move(n, s):
+    cp_schedule = SingleDiskStorageSchedule(move_data=True)
[…]
