### 2023-10-04 3fdfc13
Remove test_trivial.py -- integrated into test_validity.py
