+++ README.md
@@ -25,7 +25,10 @@ We welcome contributions to *checkpoint_schedules*!
 To contribute please consider the following steps:
 1. Fork the repository.
 2. Make your changes.
-3. Submit a pull request.
+3. Make sure that the tests pass by running `pytest test` and `pytest --nbval-lax docs/notebooks/`
+4. Add tests for your changes (if applicable).
+5. Add documentation for your changes that follows the [Sphinx](https://sphinx-rtd-tutorial.readthedocs.io/en/latest/docstrings.html) format.
+6. Submit a pull request.
 
 ## Bug reports
 Please report bugs on the [issue tracker](https://github.com/firedrakeproject/checkpoint_schedules/issues).
