+++ README.md
@@ -28,7 +28,7 @@ To contribute please consider the following steps:
 3. Make sure that the tests pass by running `pytest test` and `pytest --nbval-lax docs/notebooks/`
 4. Add tests for your changes (if applicable).
 5. Add documentation for your changes that follows the [Sphinx](https://sphinx-rtd-tutorial.readthedocs.io/en/latest/docstrings.html) format.
-6. Submit a pull request.
+6. Submit your pull request.
 
 ## Bug reports
 Please report bugs on the [issue tracker](https://github.com/firedrakeproject/checkpoint_schedules/issues).
