+++ docs/source/index.rst
@@ -19,7 +19,7 @@ If you want to quickly get up and running with *checkpoint_schedules* your adjoi
   $  pip install checkpoint-schedules
 
 * Familiarise with *checkpoint_schedules* by acessing the :ref:`introduction <introduction>`.
-* Go through to an `illustrative example <https://nbviewer.org/github/firedrakeproject/checkpoint_schedules/blob/main/docs/notebooks/tutorial.ipynb>`__ that explains the usage of *checkpoint_schedules* for step-based incremental checkpointing of the adjoints.
+* Go through to an `illustrative example <https://nbviewer.org/github/firedrakeproject/checkpoint_schedules/blob/main/docs/notebooks/tutorial.ipynb>`_ that explains the usage of *checkpoint_schedules* for step-based incremental checkpointing of the adjoints.
 
 API documentation
 =================
@@ -32,10 +32,19 @@ Contributing
 ============
 We welcome contributions to *checkpoint_schedules*!
 To contribute please consider the following steps:
+
 1. Fork the repository.
+
 2. Make your changes.
-3. Submit a pull request.
+
+3. Make sure that the tests pass by running `pytest test` and `pytest --nbval-lax docs/notebooks/`.
+
+4. Add tests for your changes (if applicable).
+
+5. Add documentation for your changes that follows the `Sphinx <https://sphinx-rtd-tutorial.readthedocs.io/en/latest/docstrings.html>`_ format.
+
+6. Submit a pull request.
 
 Bug reports
 ===========
-Please report bugs on the [issue tracker](https://github.com/firedrakeproject/checkpoint_schedules/issues).
+Please report bugs on the `issue tracker <https://github.com/firedrakeproject/checkpoint_schedules/issues>`_.
