### 2023-10-06 #5888 @Ig-dolci
Hi @diehlpk. Yes, you can start finding reviewers. We have added more information to the Readme, which now has links for package usage and installation, and pieces of information about the contribution guideline.
