+++ tests/test_multistage.py
@@ -1,19 +1,22 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
+
 import functools
 import pytest
-from checkpoint_schedules import MultistageCheckpointSchedule, \
-    Copy, Move, Forward, Reverse, EndForward, EndReverse, \
-    StorageType
+from checkpoint_schedules import (
+    MultistageCheckpointSchedule, Copy, Move, Forward, Reverse, EndForward,
+    EndReverse, StorageType)
 from checkpoint_schedules.multistage import optimal_steps_binomial
 
 
 @pytest.mark.parametrize("trajectory", ["revolve",
                                         "maximum"])
-@pytest.mark.parametrize("n, S", [(1, (0,)),
+@pytest.mark.parametrize("n, S", [
+                                  (1, (0,)),
                                   (2, (1,)),
                                   (3, (1, 2)),
-                                  (10, tuple(range(1, 10))),
+                                  (5, (2,)),
+                                  (10, tuple(range(2, 10))),
                                   (100, tuple(range(1, 100))),
                                   (250, tuple(range(25, 250, 25)))
                                   ])
@@ -25,10 +28,11 @@ def test_multistage(trajectory, n, S):
     trajectory : str
         The trajectory to use. Either `'revolve'` or `'maximum'`.
     n : int
-        The number of forward steps.
+        Number of forward steps.
     S : int
-        The number of snapshots saved in disk.
+        Number of checkpoint units.
     """
+
     @functools.singledispatch
     def action(cp_action):
         raise TypeError("Unexpected action")
@@ -36,42 +40,35 @@ def test_multistage(trajectory, n, S):
     @action.register(Forward)
     def action_forward(cp_action):
         nonlocal model_n, model_steps
-        store_ics = cp_action.write_ics
-        store_data = cp_action.write_adj_deps
-        # data.clear()
-        ics.clear()
-        # Start at the current location of the forward
-        assert model_n == cp_action.n0
-        # End at or before the end of the forward
-        assert cp_action.n1 <= n
 
-        if store_ics:
-            assert cp_action.storage == StorageType.DISK
-            assert cp_schedule.uses_storage_type(StorageType.DISK)
+        # Start at the current location of the forward
+        assert cp_action.n0 == model_n
+        # Do not advance further than the current location of the adjoint
+        assert cp_action.n1 <= n - model_r
+
+        ics.clear()
+        data.clear()
+
+        if cp_action.write_ics:
             # Advance at least one step when storing forward restart data
             assert cp_action.n1 > cp_action.n0
             # Do not advance further than one step before the current location
             # of the adjoint
             assert cp_action.n1 < n - model_r
-            # No data for these steps is stored
-            assert len(ics.intersection(range(cp_action.n0, cp_action.n1))) == 0  # noqa: E501
 
+            assert cp_action.storage == StorageType.DISK
+            assert cp_schedule.uses_storage_type(StorageType.DISK)
             ics.update(range(cp_action.n0, cp_action.n1))
-            # Written data consists of forward restart data
-            assert len(ics) > 0
-            assert len(data) == 0
-            # The checkpoint location is associated with the earliest step for
-            # which data has been stored
-            assert cp_action.n0 == min(ics)
+            assert cp_action.n0 not in snapshots
             snapshots[cp_action.n0] = (set(ics), set(data))
 
-        if store_data:
+        if cp_action.write_adj_deps:
             # Advance exactly one step when storing non-linear dependency data
             assert cp_action.n1 == cp_action.n0 + 1
             # Start from one step before the current location of the adjoint
[…]
