### 2023-10-11 d8a6eb7
Check that no working memory data is stored at the start of Copy and Move actions
