### 2023-12-14 #6148 @editorialbot
[REVIEW]: checkpoint_schedules: schedules for incremental of adjoint simulations
**Submitting author:** @Ig-dolci (<a href="http://orcid.org/0000-0002-1435-9538">Daiane Iglesia Dolci</a>)
**Repository:** https://github.com/firedrakeproject/checkpoint_schedules
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.3
