### 2024-01-05 #6148 @KYANJO
It’s going well so far about to finish

With utmost gratitude,
Brian KYANJO (He / Him / His),
President. SIAM Boise State Chapter,
VP. NSBE Boise State Society,

NSBE Public Policy SIG STEM Education, Team Co-Lead,
BSc. Physics and Mathematics | MSc. Mathematical Sciences | Computing Ph.D.
BSU,
https://github.com/KYANJO

On Fri, Jan 5, 2024 at 2:26 AM Patrick Diehl ***@***.***>
wrote:

> @matt-graham <https://github.com/matt-graham> how is your review going?
>
