### 2024-01-05 #6148 @KYANJO
I will let you know, if i get any questions

With utmost gratitude,
Brian KYANJO (He / Him / His),
President. SIAM Boise State Chapter,
VP. NSBE Boise State Society,

NSBE Public Policy SIG STEM Education, Team Co-Lead,
BSc. Physics and Mathematics | MSc. Mathematical Sciences | Computing Ph.D.
BSU,
https://github.com/KYANJO
