### 2024-01-05 firedrakeproject/checkpoint_schedules#54 @matt-graham
Referencing existing H-Revolve implementation in paper
Raising as part of https://github.com/openjournals/joss-reviews/issues/6148

The JOSS paper draft refers to a 'proof-of-concept' implementation of H-Revolve

https://github.com/firedrakeproject/checkpoint_schedules/blob/055c881297fea83f5531cfb2b6e2a33f3bc1d18f/docs/paper/paper.md?plain=1#L98-L103

From a bit of searching I came across

https://gitlab.inria.fr/adjoint-computation/H-Revolve

which I assume is what is being referred to? This should ideally be directly referenced in the JOSS paper (with citation including a URL pointing to the repository) and some explanation given to how it differs from the package here, as it appears to implement several of the same checkpoint schedule algorithms.

I also note that the three contributors on that repository are all authors on the JOSS paper, though with the exception of David Ham, do not appear to have made direct contributions to the repository here. Is the basis for including these additional authors that `checkpointing_schedules` is partly inspired or derived from this previous package? If so, again this should be clarified in the paper ideally.
