### 2024-01-05 firedrakeproject/checkpoint_schedules#55 @matt-graham
Adding an example of using the package in practice
Raising as part of JOSS review https://github.com/openjournals/joss-reviews/issues/6148

The tutorial example in https://github.com/firedrakeproject/checkpoint_schedules/blob/main/docs/notebooks/tutorial.ipynb is great at giving an understanding of the packages's API and an overview of the different checkpoint schedule algorithms implemented, however it would also be useful to have an example of how the package can be used in practice in actual adjoint computation, even if it's for a trivial problem / forward computation.

The paper indicates the package has been incoroporated in to `tlm_adjoint` (presumably once https://github.com/tlm-adjoint/tlm_adjoint/pull/382 is merged in?) - even just linking to how the package is used there as a 'next steps' tutorial example could be useful.
