### 2024-01-15 6cfa70e
Install scipy.
