### 2024-01-17 #6148 @KYANJO
I am through with my first review, and I have included comments in the pdf it's self. Find attached.
[10.21105.joss.06148.pdf](https://github.com/openjournals/joss-reviews/files/13968958/10.21105.joss.06148.pdf)

PS; If there is a better way i can upload the document please let me know, Is it better to create a pull request instead?
