+++ checkpoint_schedules/hrevolve_sequences/periodic_disk_revolve.py
@@ -1,4 +1,15 @@
-
+#!/usr/bin/python
+# Copyright (C) 2020 - 2024 Inria and Imperial College London
+# Developed originally by Guillaume Pallez (guillaume.pallez@inria.fr), Julien Herrmann (jln.herrmann@gmail.com).
+# Modified by Daiane I. Dolci (d.dolci@eimperial.ic.ac.uk) and David A. Ham (david.ham@imperial.ac.uk).
+# This program is free software: you can redistribute it and/or modify it
+# under the terms of the GNU General Public License as published by the
+# Free Software Foundation, either version 3 of the License, or (at your option)
+# any later version. This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+# or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+# You should have received a copy of the GNU General Public License along with
+# this program. If not, see <https://www.gnu.org/licenses/>.
 """This module contains the periodic disk revolve checkpoint schedule."""
 from functools import partial
 from .basic_functions import (Operation as Op, Sequence, Function, beta)
