+++ checkpoint_schedules/multistage.py
@@ -1,3 +1,16 @@
+#!/usr/bin/python
+# Copyright (C) 2020 - 2024 The University of Edinburgh and Imperial College London
+# Developed originally by James R. Maddison (j.r.maddison@ed.ac.uk).
+# Modified by Daiane I. Dolci (d.dolci@eimperial.ic.ac.uk) and David A. Ham (david.ham@imperial.ac.uk).
+# This program is free software: you can redistribute it and/or modify it
+# under the terms of the GNU General Public License as published by the
+# Free Software Foundation, either version 3 of the License, or (at your option)
+# any later version. This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+# or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+# You should have received a copy of the GNU General Public License along with
+# this program. If not, see <https://www.gnu.org/licenses/>.
+
 import functools
 from operator import itemgetter
 from .schedule import CheckpointSchedule, Forward, Reverse, Copy, Move, \
