### 2024-01-23 73d642c
Small modifications related to JOSS rewiew.
