### 2024-01-23 5bc8feb
Add RAM meaning.
