+++ checkpoint_schedules/mixed.py
@@ -1,12 +1,5 @@
+# SPDX-License-Identifier: LGPL-3.0-only
 # Copyright (C) 2020 - 2024 The University of Edinburgh and Imperial College London
-# This program is free software: you can redistribute it and/or modify it
-# under the terms of the GNU General Public License as published by the
-# Free Software Foundation, either version 3 of the License, or (at your option)
-# any later version. This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
-# or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
-# You should have received a copy of the GNU General Public License along with
-# this program. If not, see <https://www.gnu.org/licenses/>.
 # Developed originally by James R. Maddison (j.r.maddison@ed.ac.uk).
 # Modified by Daiane I. Dolci (d.dolci@eimperial.ic.ac.uk) and David A. Ham (david.ham@imperial.ac.uk).
 import warnings
