+++ checkpoint_schedules/hrevolve.py
@@ -1,5 +1,6 @@
 # SPDX-License-Identifier: LGPL-3.0-only
-# Copyright (C) 2020 - 2024 The University of Edinburgh and Imperial College London
+# Copyright (C) 2020 - 2024 The University of Edinburgh and Imperial College
+# London
 # Developed by Daiane I. Dolci (d.dolci@eimperial.ic.ac.uk),
 # James R. Maddison (j.r.maddison@ed.ac.uk)
 # and David A. Ham (david.ham@imperial.ac.uk).
