### 2024-01-24 99e5fd1
Update docs/paper/paper.md
Co-authored-by: David A. Ham <david.ham@imperial.ac.uk>
