### 2024-01-29 8b6c4cd
Update docs/source/index.rst
Co-authored-by: James R. Maddison <j.r.maddison@ed.ac.uk>
