### 2024-01-31 245a1dc
Dolci/adding copyright (#58)
Add spdx copyright headers and authorship information to source files.
