+++ checkpoint_schedules/twolevel_binomial.py
@@ -1,3 +1,9 @@
+# SPDX-License-Identifier: LGPL-3.0-only
+# Copyright (C) 2020 - 2024 The University of Edinburgh and Imperial College
+# London
+# Developed originally by James R. Maddison (j.r.maddison@ed.ac.uk).
+# Modified by Daiane I. Dolci (d.dolci@eimperial.ic.ac.uk)
+# and David A. Ham (david.ham@imperial.ac.uk).
 from .schedule import CheckpointSchedule, Forward, Reverse, Copy, Move, \
     EndForward, EndReverse, StorageType
 from .multistage import n_advance
