+++ pyproject.toml
@@ -7,7 +7,7 @@ build-backend = "setuptools.build_meta"
 name = "checkpoint_schedules"
 version = "1.0.0"
 authors = [
-  { name="Daiane I. Dolci ", email="ddolci@eimperial.ic.ac.uk" },
+  { name="Daiane I. Dolci ", email="d.dolci@eimperial.ic.ac.uk" },
   { name="James R. Maddison", email="j.r.maddison@ed.ac.uk" },
   { name="David A. Ham", email="david.ham@imperial.ac.uk" },
   { name="Guillaume Pallez", email="guillaume.pallez@inria.fr" },
