### 2024-02-07 21bec1c
Notebook containing an application of checkpoint_schedules package.
