### 2024-02-12 3c8cf25
add sympy
