+++ .github/workflows/build.yml
@@ -27,7 +27,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install pytest numpy nbval tabulate scipy
+        python -m pip install pytest numpy nbval tabulate scipy sympy
         python -m pip install -e .
     - name: Test with pytest
       run: |
