+++ docs/paper/paper.bib
@@ -61,7 +61,7 @@
 
 @article{herrmann2020,
   title = {H-Revolve: a framework for adjoint computation on synchronous hierarchical platforms},
-  author = {Herrmann, Julien},
+  author = {Herrmann, Julien and Pallez, Guillaume},
   journal = {ACM Transactions on Mathematical Software (TOMS)},
   volume = {46},
   number = {2},
