### 2024-03-01 2c140e2
Dolci/schedules application (#56)
* Notebook containing an application of checkpoint_schedules package.
---------

Co-authored-by: James R. Maddison <j.r.maddison@ed.ac.uk>
