### 2024-03-01 ce51c90
Minor paper modifications (#57)
* Small corrections related to the JOSS review.
---------

Co-authored-by: David A. Ham <david.ham@imperial.ac.uk>
