### 2024-03-01 firedrakeproject/checkpoint_schedules#54 @Ig-dolci
@matt-graham

`checkpoint_schedules` does indeed incorporate code from the reference implementation presented in Hermann & Pallez (2020), though several changes and fixes were required in order to make the code production-ready. Hermann and Pallez contributed to those fixes, though they did not personally make commits to the new repository. We have made the contributions of all the authors more explicit by introducing an author contribution section in the paper, and by adding copyright headers with accurate authorship information to all of the source files. We have not added a direct link to the previous git repository, because the canonical location for that code is now here and because the long-term status of the other repository is uncertain. That repository referenced from Hermann & Pallez (2020) as is appropriate because that version of the software was used in that paper.

There are also pull requests that go along with that:
- [PR 58](https://github.com/firedrakeproject/checkpoint_schedules/pull/58) adds copyright headers with accurate authorship information to all of the source files.
-  [PR 61](https://github.com/firedrakeproject/checkpoint_schedules/pull/61) creates the author contribution section in the paper.
