### 2024-03-01 firedrakeproject/checkpoint_schedules#55 @Ig-dolci
@matt-graham We appreciate your suggestion to include an example application in our documentation.

Following your suggestion and seeking to make the usage of `checkpoint_schedules` more accessible, we have included an example of checkpoint_schedules usage in an adjoint-based gradient problem. You can access it in the following notebook: [checkpoint_schedule application: Adjoint-Based Gradient with the Burger's Equation](https://github.com/firedrakeproject/checkpoint_schedules/blob/main/docs/notebooks/burger.ipynb).

We are also pleased to inform you that checkpoint_schedules is now incorporated into the pyadjoint package (PR [121](https://github.com/dolfin-adjoint/pyadjoint/pull/121)).
