### 2024-03-08 firedrakeproject/checkpoint_schedules#54 @matt-graham
@Ig-dolci - thanks for clarifying and adding the author contribution section to paper.
