### 2024-03-08 firedrakeproject/checkpoint_schedules#55 @matt-graham
@Ig-dolci - thanks the new notebook is fantastic!
