+++ CITATION.cff
@@ -1,0 +1,21 @@
+cff-version: 1.1.0
+message: "If you use this software, please cite it as below."
+authors:
+- family-names: Dolci
+  given-names: Daiane I.
+  orcid: https://orcid.org/0000-0002-1435-9538
+- family-names: Maddison
+  given-names: James R.
+  orcid: https://orcid.org/0000-0001-5742-4363
+- family-names: Ham
+  given-names: David A.
+  orcid: https://orcid.org/0000-0001-9545-9110
+- family-names: Pallez
+  given-names: Guillaume
+  orcid: https://orcid.org/0000-0001-8862-3277
+- family-names: Herrmann
+  given-names: Julien
+  orcid: https://orcid.org/0000-0003-4935-2368
+title: "checkpoint_schedules: schedules for incremental checkpointing of adjoint simulations"
+version: v1.0.3
+date-released: 2023-03-14
