### 2024-03-14 #6148 @Ig-dolci
> @Ig-dolci Please finish these tasks, so we can proceed with the paper
>
> * [x]      Double check authors and affiliations (including ORCIDs)
> * [x]      Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> * [x]   Archive the release on Zenodo/figshare/etc and post the DOI here.
> * [x]    Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * [x]  Make sure that the license listed for the archive is the same as the software license.

@diehlpk below the latest release and Zenodo DOI.
Release version: [v1.0.3](https://github.com/firedrakeproject/checkpoint_schedules/releases/tag/v1.0.3)
Zenodo DOI: [10817312](https://zenodo.org/records/10817312)
