### 2024-03-14 #6148 @danielskatz
@Ig-dolci - I'm the track editor who will finish the processing of this submission.

I've suggested some small changes in https://github.com/firedrakeproject/checkpoint_schedules/pull/65. Please merge this, or let me know what you disagree with.

In addition, please add countries to all of the affiliations, though we don't need mailing address / postal codes. Usually, institution & country is enough, though department or similar is also fine.

Also in the paper, "in the frame of DASH" in the acknowledgements sounds awkward to me, but I'm not sure how to fix it, as I don't really know what this means.

Finally, in the references, the one for Pringle seems to be missing information about how to access it. Should it have a URL or something else?
