### 2024-03-15 01e52b7
Patch 1 (#65)
* Small changes in paper.md and paper.bib.
