+++ docs/paper/paper.bib
@@ -176,7 +176,7 @@ eprint = {
   number = {38}, 
   pages = {1292}, 
   author = {Sebastian K. Mitusch and Simon W. Funke and Jørgen S. Dokken}, 
-  title = {dolfin-adjoint 2018.1: automated adjoints for FEniCS and Firedrake}, 
+  title = {dolfin-adjoint 2018.1: automated adjoints for FEniCS and {F}iredrake}, 
   journal = {Journal of Open Source Software} 
 }
 
