+++ docs/paper/paper.md
@@ -33,7 +33,7 @@ date:  September 2023
 bibliography: paper.bib
 ---
 # Summary
-*checkpoint_schedules* provides schedules for step based incremental
+*checkpoint_schedules* provides schedules for step-based incremental
 checkpointing of the adjoints to computer models. The schedules contain
 instructions indicating the sequence of forward and adjoint steps to be
 executed, and the data storage and retrieval to be performed. These
@@ -52,13 +52,13 @@ The use of adjoint calculations to compute the gradient of a quantity of
 interest resulting from the solution of a system of partial differential
 equations (PDEs) is widespread and well-established. The resulting gradient may
 be employed for many purposes, including topology
-optimisation [@papadopoulos2021computing], inverse problems [@Plessix2006],
+optimisation [@papadopoulos2021computing], inverse problems [@Plessix2006], and
 flow control [@Jansen2011]. 
 
 Solving the adjoint to a non-linear time-dependent PDE requires the forward PDE
 to be solved first. The adjoint PDE is then solved in a reverse time
 order, but depends on the forward state. Storing the entire forward state in
-preparation for the adjoint calculation has a memory footprint linear in the
+preparation for the adjoint calculation has a memory footprint that is linear in the
 number of time steps. For sufficiently large problems this will exhaust the
 memory of any computer system. 
 
@@ -68,21 +68,21 @@ calculation progresses, the forward computation is progressively rerun from the
 latest available stored state up to the current adjoint step. This enables less
 forward state to be stored, at the expense of a higher computational cost as
 forward steps are run more than once. @griewank2000algorithm proposed a
-checkpointing algorithm which is optimal under certain assumptions, including
+checkpointing algorithm, which is optimal under certain assumptions, including
 that the number of steps is known in advance, and that all the storage has
 equal access cost. Subsequent authors have produced checkpointing algorithms
 that relax these requirements in various ways, such as by accounting for
-different types of storage (e.g. memory and disk) or by not
+different types of storage (e.g., memory and disk) or by not
 requiring the number of steps to be known in advance, for example 
-[@stumm2009multistage; @aupy2016optimal; @schanen2016; @aupy2017periodicity; 
-@herrmann2020; @maddison2023; @Zhang_2023]. 
+@stumm2009multistage, @aupy2016optimal, @schanen2016, @aupy2017periodicity, 
+@herrmann2020, @maddison2023, and @Zhang_2023. 
 
 # Statement of need
 
 This situation is typical across computational mathematics: there exists a
 diversity of algorithms whose applicability and optimality depends on the
 nature and parameters of the problem to be solved. From the perspective of
-users who wish to construct adjoint solvers this creates the need to swap out
+users who wish to construct adjoint solvers, this creates the need to swap out
 different checkpointing algorithms in response to changes in the equations,
 discretisations, and computer systems with which they work. Those users will
 often lack the expertise or the time to continually reimplement additional
@@ -90,12 +90,12 @@ algorithms in their framework. Further, such reimplementation work is wasteful
 and error-prone. 
 
 *checkpointing_schedules* provides a number of checkpointing algorithms
-accessible through a common interface and these are interchangeable without
+accessible through a common interface, and these are interchangeable without
 recoding. This can be used in conjunction with an adjoint framework such as
 tlm_adjoint or pyadjoint and a compatible PDE framework, such as Firedrake
-[@FiredrakeUserManual] or FEniCS [@AlnaesEtal2015] to enable users to create
+[@FiredrakeUserManual] or FEniCS [@AlnaesEtal2015], to enable users to create
 adjoint solvers for their choice of PDE, numerical methods, and checkpointing
[…]
