+++ docs/paper/paper.bib
@@ -141,6 +141,7 @@ eprint = {
   author = {Pringle, GC and Jones, Daniel C and Goswami, Sudipta and Narayanan, Sri Hari Krishna and Goldberg, Daniel},
   year = {2016},
   publisher = {EPCC},
+  url = {www.archer.ac.uk/community/eCSE/eCSE03-09/eCSE03-09_White_Paper.pdf},
 }
 
 @article{goldberg2020bathymetric,
