### 2024-03-15 #6148 @Ig-dolci
@danielskatz Your pull request [PR 65](https://github.com/firedrakeproject/checkpoint_schedules/pull/65) has been merged.

In addition, I have opened [PR 66](https://github.com/firedrakeproject/checkpoint_schedules/pull/66), which includes the countries of all affiliations and a URL was added in the Pringle reference. Could you please verify if it sounds good for the final JOSS paper?

The mention of "in the frame of DASH" in the acknowledgements section refers to a funding project. I verified this terminology is usual in other papers authored by one of our authors. So, we should keep it in this form.
