### 2024-03-15 #6148 @danielskatz
Thanks - this looks good, though I'm not sure if the type for Pringle should be article or misc. We can try this and see how it looks.
