+++ docs/paper/paper.bib
@@ -136,12 +136,12 @@ eprint = {
   doi = {10.1080/10556788.2016.1230612},
 }
 
-@article{pringle2016providing,
+@misc{pringle2016providing,
   title = {Providing the ARCHER community with adjoint modelling tools for high-performance oceanographic and cryospheric computation},
   author = {Pringle, GC and Jones, Daniel C and Goswami, Sudipta and Narayanan, Sri Hari Krishna and Goldberg, Daniel},
   year = {2016},
   publisher = {EPCC},
-  url = {www.archer.ac.uk/community/eCSE/eCSE03-09/eCSE03-09_White_Paper.pdf},
+  url = {www.archer.ac.uk/community/eCSE/eCSE03-09/eCSE03-09\_White_Paper.pdf},
 }
 
 @article{goldberg2020bathymetric,
