### 2024-03-15 #6148 @Ig-dolci
I think misc is better for this case. I just changed from article to misc.
