+++ docs/paper/paper.bib
@@ -136,7 +136,7 @@ eprint = {
   doi = {10.1080/10556788.2016.1230612},
 }
 
-@article{pringle2016providing,
+@misc{pringle2016providing,
   title = {Providing the ARCHER community with adjoint modelling tools for high-performance oceanographic and cryospheric computation},
   author = {Pringle, GC and Jones, Daniel C and Goswami, Sudipta and Narayanan, Sri Hari Krishna and Goldberg, Daniel},
   year = {2016},
