### 2024-03-19 e6106b0
Small changes (#66)
* Move the reference to misc and add a link to this reference in the paper.bib
* Add the country and city associated with the affiliations.
