+++ docs/paper/paper.bib
@@ -136,11 +136,12 @@ eprint = {
   doi = {10.1080/10556788.2016.1230612},
 }
 
-@article{pringle2016providing,
+@misc{pringle2016providing,
   title = {Providing the ARCHER community with adjoint modelling tools for high-performance oceanographic and cryospheric computation},
   author = {Pringle, GC and Jones, Daniel C and Goswami, Sudipta and Narayanan, Sri Hari Krishna and Goldberg, Daniel},
   year = {2016},
   publisher = {EPCC},
+  url = {https://nora.nerc.ac.uk/id/eprint/516314/},
 }
 
 @article{goldberg2020bathymetric,
