### 2024-03-22 #6148 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
